+++ css/style.css
@@ -54,6 +54,7 @@ body {
 .bg-album {
   position: relative;
   background-color: #f8f9fa;
+  min-height: 600px;
 }
 .col {
   max-width: 360px;
@@ -172,7 +173,7 @@ body {
   align-items: center;
   justify-content: center;
   height: 318px;
-  width: 328px;
+  width: 280px;
   transition-duration: 0.6s;
   padding: 15px;
   border-radius: 7%;
@@ -268,6 +269,7 @@ li {
   justify-content: flex-start;
   padding: 20px 30px;
   transition-duration: 0.5s;
+  z-index: 100000;
 }
 .close{
   position: absolute;
@@ -323,7 +325,7 @@ h4{
   width: 100%;
   height: 100%;
   background: rgba(0, 0, 0, .1);
-  z-index: 1000;
+  z-index: 100;
 }
 .hearts-container {
   display: flex;
@@ -372,6 +374,7 @@ th{
   text-align: center;
   margin: 15px auto;
   border-radius: 10%;
+  color: #000;
 }
 .btn-group {
   width: 400px;
