+++ js/pages/statisticsPage/statisticFilter.js
@@ -1,121 +1,158 @@
-import {createStatistics} from './statisticPageLayout';
+import { createRow } from '../pagesLayout/statisticLayout';
 
 function cleanTBody() {
-    let tbody = document.querySelector('tbody');
-    while(tbody.firstChild) {
-        tbody.firstChild.remove();
-    }
+  const tbody = document.querySelector('tbody');
+  while (tbody.firstChild) {
+    tbody.firstChild.remove();
+  }
+}
+
+function updateStatisticContent(num, category, word, translation, trainClick, playClick, errors) {
+  const tbody = document.querySelector('tbody');
+  let percent = Math.floor((errors / playClick) * 100);
+  if (!percent) {
+    percent = 0;
+  }
+  const row = document.createElement('tr');
+  createRow(row, num);
+  createRow(row, (category.charAt(0).toUpperCase() + category.slice(1)));
+  createRow(row, word);
+  createRow(row, translation);
+  createRow(row, trainClick, 'train-column');
+  createRow(row, playClick, 'play-column');
+  createRow(row, errors, 'errors-column');
+  createRow(row, percent, 'percent-column');
+  tbody.appendChild(row);
 }
 
 function returnAllWords() {
-    let statistic = JSON.parse(localStorage.getItem('statistic'));
-    const keys = Object.keys(statistic);
+  const statistic = JSON.parse(localStorage.getItem('statistic'));
+  const keys = Object.keys(statistic);
 
-    let arr = [];
-    keys.forEach(key => {
-        if (key !== 'sections') {
-            statistic[key].forEach(i => {
-                arr.push(i);
-            })
-        }
-    })
-    return(arr);
+  const arr = [];
+  keys.forEach((key) => {
+    if (key !== 'sections') {
+      statistic[key].forEach((i) => {
+        arr.push(i);
+      });
+    }
+  });
+  return (arr);
 }
 
 function tableSort(arr, filter) {
-    arr.sort((a, b) => {
-        let aNew;
-        let bNew;
-        if (filter === 'percent') {
-            aNew = Math.floor(a.errors/a.playClick * 100) | 0;
-            bNew = Math.floor(b.errors/b.playClick * 100) | 0;
-        } else {
-            aNew = a[filter];
-            bNew = b[filter];
-        }
-        if (aNew > bNew) {
-            return 1;
-        } else {
-            return -1;
-        }
-    })
-    return arr;
+  arr.sort((a, b) => {
+    let aNew;
+    let bNew;
+    if (filter === 'percent') {
+      aNew = Math.floor((a.errors / a.playClick) * 100);
+      bNew = Math.floor((b.errors / b.playClick) * 100);
+      if (!aNew) {
+        aNew = 0;
+      }
+      if (!bNew) {
+        bNew = 0;
+      }
+    } else {
+      aNew = a[filter];
+      bNew = b[filter];
+    }
+    if (aNew > bNew) {
+      return 1;
+    }
+    return -1;
+  });
+  return arr;
 }
+function createStatisticsBySection(sections) {
+  let num = 1;
 
-function createStatisticsBySection(keys) {
-
-    let statistic = JSON.parse(localStorage.getItem('statistic'));
-    let num = 1;
-
-    keys.forEach(key => {
-        if (key !== 'sections') {
-            statistic[key].forEach(({title, translate, trainClick, playClick, errors}) => {
-                createStatistics(num, key, title, translate, trainClick, playClick, errors);
-                num++;
-            })
-        }
-    })
+  const newData = JSON.parse(localStorage.getItem('statistic'));
+  sections.forEach((section) => {
+    if (section !== 'sections') {
+      newData[section].forEach(({
+        title, translate, trainClick, playClick, errors, key,
+      }) => {
+        updateStatisticContent(num, key, title, translate, trainClick, playClick, errors);
+        num += 1;
+      });
+    }
+  });
 }
 
 function createStatisticsByWord(arr) {
-    let num = 1;
+  let num = 1;
 
-    arr.forEach(({title, translate, trainClick, playClick, errors, key}) => {
-        createStatistics(num, key, title, translate, trainClick, playClick, errors);
-        num++;
-    })
+  arr.forEach(({
+    title, translate, trainClick, playClick, errors, key,
+  }) => {
+    updateStatisticContent(num, key, title, translate, trainClick, playClick, errors);
+    num += 1;
+  });
 }
-
+function sortAllWords(title, reverse) {
+  if (reverse) {
+    createStatisticsByWord(tableSort(returnAllWords(), title));
+  } else {
+    createStatisticsByWord(tableSort(returnAllWords(), title).reverse());
+  }
+}
 function statisticFilter(filter, reverse) {
-    cleanTBody();
+  cleanTBody();
 
-    let statistic = JSON.parse(localStorage.getItem('statistic'));
-    const keys = Object.keys(statistic);
+  const statistic = JSON.parse(localStorage.getItem('statistic'));
+  const keys = Object.keys(statistic);
 
-    switch(filter) {
-        case "Category":
-            reverse ? keys.sort() : keys.sort().reverse();
-            createStatisticsBySection(keys);
-            break;
-        case "Sr. No.":
-            reverse ? keys : keys.reverse();
-            createStatisticsBySection(keys);
-            break;
-        case "Word":
-            let newArr = (reverse ? tableSort(returnAllWords(), 'title') : tableSort(returnAllWords(), 'title').reverse());
-            createStatisticsByWord(newArr);
-            break;
-        case "Translation":
-            let arr = (reverse ? tableSort(returnAllWords(), 'translate') : tableSort(returnAllWords(), 'translate').reverse());
-            createStatisticsByWord(arr);
-            break;
-        case "trainClick":
-            let clickArr = (reverse ? tableSort(returnAllWords(), 'trainClick') : tableSort(returnAllWords(), 'trainClick').reverse());
-            createStatisticsByWord(clickArr);
-            break;
-        case "playClick":
-            let clickArrPlay = (reverse ? tableSort(returnAllWords(), 'playClick') : tableSort(returnAllWords(), 'playClick').reverse());
-            createStatisticsByWord(clickArrPlay);
-            break;
-        case "errors":
-            let errorClick = (reverse ? tableSort(returnAllWords(), 'errors') : tableSort(returnAllWords(), 'errors').reverse());
-            createStatisticsByWord(errorClick);
-            break;
-        case "percent":
-            let perClick = (reverse ? tableSort(returnAllWords(), "percent") : tableSort(returnAllWords(), "percent").reverse());
-            createStatisticsByWord(perClick);
-            break;
-    }
+  switch (filter) {
+    case 'Category':
+      if (reverse) {
+        createStatisticsBySection(keys.sort());
+      } else {
+        createStatisticsBySection(keys.sort().reverse());
+      }
+      break;
+    case 'Sr.No.':
+      if (reverse) {
+        createStatisticsBySection(keys);
+      } else {
+        createStatisticsBySection(keys.reverse());
+      }
+      break;
+    case 'Word':
+      sortAllWords('title', reverse);
+      break;
+    case 'Translation':
+      sortAllWords('translate', reverse);
+      break;
+    case 'trainClick':
+      sortAllWords('trainClick', reverse);
+      break;
+    case 'playClick':
+      sortAllWords('playClick', reverse);
+      break;
+    case 'errors':
+      sortAllWords('errors', reverse);
+      break;
+    case 'percent':
+      sortAllWords('persent', reverse);
+      break;
+    default:
+      break;
+  }
 }
 
 function returnEightErrorsOrLess() {
-    let errorClick = tableSort(returnAllWords(), 'errors');
+  let errorClick = tableSort(returnAllWords(), 'errors');
 
-    errorClick = errorClick.filter(i => i.errors !== 0).reverse();
-    errorClick.length >= 8 ? errorClick = errorClick.slice(0, 8) : errorClick = errorClick.slice(0, errorClick.length);
+  errorClick = errorClick.filter((i) => i.errors !== 0).reverse();
+  if (errorClick.length >= 8) {
+    errorClick = errorClick.slice(0, 8);
+  } else {
+    errorClick = errorClick.slice(0, errorClick.length);
+  }
 
-    return errorClick;
+  return errorClick;
 }
 
 export default statisticFilter;
-export {returnEightErrorsOrLess, returnAllWords};
+export { returnEightErrorsOrLess, returnAllWords, updateStatisticContent };
