+++ package.json
@@ -3,6 +3,9 @@
     "@babel/core": "^7.20.12",
     "@babel/preset-env": "^7.20.2",
     "babel-loader": "^9.1.2",
+    "eslint": "^8.33.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.27.5",
     "json-server": "^0.17.1",
     "normalize.css": "^8.0.1",
     "webpack": "^5.75.0",
